src/requirements.txt
bcrypt
ifcfg
psutil
pyotp
Flask
flask-cors
icmplib
gunicorn
requests
tcconfig
sqlalchemy
sqlalchemy_utils
psycopg2
mysqlclient
tzlocal
python-jose